initial moving around of files
