+++ @types/index.d.ts
@@ -4,18 +4,18 @@ import { CustomError } from "./objects/error";
 import { CustomPromise } from "./objects/promise";
 
 // classes
-import { AwayFromKeyboard } from "./classes/afk";
-import { Banphrase } from "./classes/banphrase";
-import { Channel } from "./classes/channel";
-import { ChatModule } from "./classes/chat-module";
-import { Command } from "./classes/command";
+import { AwayFromKeyboard } from "./supibot/afk";
+import { Banphrase } from "./supibot/banphrase";
+import { Channel } from "./supibot/channel";
+import { ChatModule } from "./supibot/chat-module";
+import { Command } from "./supibot/command";
 import { Cron } from "./classes/cron";
 import { Config } from "./classes/config";
-import { Filter } from "./classes/filter";
+import { Filter } from "./supibot/filter";
 import { GotProxy as Got } from "./classes/got";
-import { Platform } from "./classes/platform";
-import { Reminder } from "./classes/reminder";
-import { User } from "./classes/user";
+import { Platform } from "./supibot/platform";
+import { Reminder } from "./supibot/reminder";
+import { User } from "./supibot/user";
 
 // singletons
 import { CacheSingleton } from "./singletons/cache";
