+++ @types/supibot/platforms/twitch-platform.d.ts
