+++ supibot/banphrase.js
