+++ supibot/chat-module.js
